Quick Code Cleanup *Move TagBlock.h to DataTypes
