--- xlive/Blam/Cache/Cache.h
@@ -1,4 +1,0 @@
-#pragma once
-#include "DataTypes\DataTypes.h"
-#include "TagBlocks\TagBlocks.h"
-#include "TagGroups\TagGroups.h"
