--- xlive/Blam/Cache/TagBlocks/TagBlocks.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "Blam\Cache\TagBlocks\char.h"
