--- xlive/Blam/Cache/TagGroups/TagGroups.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "char.hpp"
