+++ xlive/Blam/Cache/TagGroups/antenna_definition.hpp
@@ -4,7 +4,7 @@
 #pragma pack(push,1)
 #include "..\Blam\Cache\DataTypes\DataTypes.h"
 #include "..\Blam\Cache\TagGroups.hpp"
-#include "..\Blam\Cache\TagBlocks\TagBlock.h"
+#include "..\Blam\Cache\DataTypes\TagBlock.h"
 #include "..\Blam\Maths\real_math.hpp"
 /*********************************************************************
 * name: antenna
