+++ xlive/H2MOD/Modules/CustomMenu/CustomMenu.cpp
@@ -12,6 +12,7 @@
 #include "H2MOD\Modules\Updater\Updater.h"
 #include "H2MOD\Modules\Config\Config.h"
 #include "H2MOD\Modules\Networking\NetworkSession\NetworkSession.h"
+#include "..\H2MOD\Tags\TagInterface.h"
 
 extern DWORD H2BaseAddr;
 extern bool H2IsDediServer;
