+++ xlive/H2MOD/Variants/DeviceShop/DeviceShop.cpp
@@ -2,6 +2,7 @@
 #include <unordered_map>
 #include "Globals.h"
 #include "DeviceShop.h"
+#include "..\H2MOD\Tags\TagInterface.h"
 extern void __cdecl print_to_console(char *output);
 extern void GivePlayerWeaponDatum(datum unit_datum, datum weapon_datum);
