+++ xlive/H2MOD/Variants/H2Final/H2Final.cpp
@@ -1,6 +1,7 @@
 #include "stdafx.h"
 #include "Globals.h"
 #include "H2MOD\Modules\Networking\NetworkSession\NetworkSession.h"
+#include "..\H2MOD\Tags\TagInterface.h"
 
 /*
 Todo:
