+++ xlive/H2MOD/Variants/H2X/H2X.cpp
@@ -1,5 +1,6 @@
 #include "stdafx.h"
 #include "H2MOD\Variants\H2X\H2X.h"
+#include "..\H2MOD.h"
 
 #include "H2MOD\Tags\TagInterface.h"
